refactor: Move author interfaces and utility functions into common files
